System::Context: replaced done queue with command property in Result;¹ tidy & doc tidy.
1: Without this, there's pointless removal of the command unit
from the done queue when a command completes, because clients
want to handle the result of the execution immediately, without
other commands trampling the state they're somewhat expecting.
The thought is a bit weak when you consider other commands may
be first in line, with a different completion order than you
might've wanted (due to the number of stages required), but it
also introduces pointless buffering in the general case. If the
user needs buffering, they can do it on their own, and context
is no worse off for it. This is a good reminder that explicit
execution is the only way a command can be completed (i.e.,
there're no hidden mechanics.. not counting termination stuff).